Re-organize file and folder structure
