+++ src/interfaces/coral_web/src/components/Conversation/Composer/DragDropFileUploadOverlay.tsx
